Patch to fix Multi Submit Checkout
When WorldpayJsonCw is used with OneStepCheckout, the method to disable to the checkout being submitted more than once is no longer effective due to browser CORS policy. This patch fixes behaviour so "proceed to order" button is disabled and submitting order progress status is shown as per a native Payment Method.